requirements.txt
Flask==2.3.3
geopy==2.4.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
